+++ jest/jestSetup.js
@@ -53,6 +53,18 @@ jest.mock('react-native', () => {
  * before mocking it here.
  */
 
+// As instructed at https://reactnavigation.org/docs/testing/.
+jest.mock('react-native-reanimated', () => {
+  /* eslint-disable-next-line global-require */
+  const Reanimated = require('react-native-reanimated/mock');
+
+  // The mock for `call` immediately calls the callback which is incorrect
+  // So we override it with a no-op
+  Reanimated.default.call = () => {};
+
+  return Reanimated;
+});
+
 jest.mock('@react-native-community/async-storage', () => mockAsyncStorage);
 
 jest.mock('react-native-sound', () => () => ({
