+++ package.json
@@ -34,9 +34,9 @@
     "@expo/react-native-action-sheet": "^3.4.0",
     "@react-native-community/async-storage": "^1.6.3",
     "@react-native-community/cameraroll": "^1.7.2",
+    "@react-native-community/masked-view": "^0.1.10",
     "@react-native-community/netinfo": "^5.9.5",
-    "@react-navigation/core": "^3.5.0",
-    "@react-navigation/native": "^3.6.2",
+    "@react-navigation/core": "^3.7.6",
     "@sentry/react-native": "^1.0.9",
     "@unimodules/core": "~5.1.2",
     "@zulip/shared": "^0.0.2",
@@ -66,9 +66,11 @@
     "react-native-image-picker": "^2.3.3",
     "react-native-notifications": "^1.2.0",
     "react-native-photo-view": "alwx/react-native-photo-view#c58fd6b30",
+    "react-native-reanimated": "^1.0.0",
     "react-native-safari-view": "2.0.0",
     "react-native-safe-area": "^0.4.1",
-    "react-native-screens": "^1.0.0-alpha.23",
+    "react-native-safe-area-context": "^3.1.7",
+    "react-native-screens": "^2.10.1",
     "react-native-simple-toast": "^1.0.0",
     "react-native-sound": "^0.11.0",
     "react-native-text-input-reset": "^1.0.2",
@@ -76,11 +78,11 @@
     "react-native-url-polyfill": "1.2.0-rc.0",
     "react-native-vector-icons": "^6.6.0",
     "react-native-webview": "~10.0.0",
-    "react-navigation": "^3.13.0",
-    "react-navigation-drawer": "~1.4.0",
+    "react-navigation": "^4.4.0",
+    "react-navigation-drawer": "^2.5.0",
     "react-navigation-redux-helpers": "^3.0.8",
-    "react-navigation-stack": "1.5.3",
-    "react-navigation-tabs": "~1.2.0",
+    "react-navigation-stack": "^2.8.2",
+    "react-navigation-tabs": "^2.9.0",
     "react-redux": "^5.0.7",
     "redux": "^4.0.0",
     "redux-action-buffer": "^1.2.0",
