+++ src/nav/AppNavigator.js
@@ -1,5 +1,5 @@
 /* @flow strict-local */
-import { createStackNavigator } from 'react-navigation';
+import { createStackNavigator } from 'react-navigation-stack';
 
 import AccountPickScreen from '../account/AccountPickScreen';
 import RealmScreen from '../start/RealmScreen';
